+++ src/main/java/roomescape/reservation/controller/ReservationController.java
@@ -78,29 +78,9 @@ public class ReservationController {
                 .body(reservationCreateResponse);
     }
 
-    @PostMapping("/waiting")
-    public ResponseEntity<ReservationResponse> createWaitingReservation(ReservationCreateRequest request) {
-        Long detailId = reservationDetailService.findReservationDetailId(request);
-        ReservationRequest reservationRequest = new ReservationRequest(request.memberId(), detailId);
-
-        reservationService.findReservationByDetailId(reservationRequest);
-        ReservationResponse reservationCreateResponse = waitingService.addReservationWaiting(reservationRequest);
-
-        URI uri = URI.create("/reservations/waiting/" + reservationCreateResponse.id());
-        return ResponseEntity.created(uri)
-                .body(reservationCreateResponse);
-    }
-
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteReservation(@PathVariable long id) {
         reservationService.deleteReservation(id);
-        return ResponseEntity.noContent()
-                .build();
-    }
-
-    @DeleteMapping("/waiting/{id}")
-    public ResponseEntity<Void> deleteReservationWaiting(@PathVariable long id) {
-        waitingService.removeReservations(id);
         return ResponseEntity.noContent()
                 .build();
     }
