+++ src/main/resources/static/js/reservation-mine.js
@@ -49,7 +49,7 @@ function render(data) {
 }
 
 function requestDeleteWaiting(id) {
-    const endpoint = `/reservations/waiting/${id}`;
+    const endpoint = `/waiting-reservations/${id}`;
     return fetch(endpoint, {
         method: 'DELETE'
     }).then(response => {
